requirements.txt
streamlit
PySide6
requests
python-dotenv